# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: MAX Together Backend API
  version: 1.0.0-mvp
  description: Финальный контракт MVP для персонального поиска событий и совместного выбора в комнате из двух участников в
    MAX.
servers:
- url: https://worknet.team/api/v1
  description: Production
- url: http://localhost:8080/api/v1
  description: Local
...
security:
- bearerAuth: []
paths:
...
  /feed/home:
    get:
      tags:
      - Discovery
      operationId: getHomeFeed
      summary: Получить все секции домашней ленты
      parameters:
      - $ref: '#/components/parameters/CityId'
      - name: lat
        in: query
        schema:
          type: number
          minimum: -90
          maximum: 90
        description: Необязательная координата для секции nearby; передаётся только вместе с lng и после согласия пользователя
      - name: lng
        in: query
        schema:
          type: number
          minimum: -180
          maximum: 180
        description: Необязательная координата для секции nearby; передаётся только вместе с lat и после согласия пользователя
      - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Render-ready лента
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeFeedResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
...
  /behavior/events:batch:
    post:
      tags:
      - Behavior
      operationId: ingestBehaviorEvents
      summary: Идемпотентно принять пачку рекомендательных сигналов
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BehaviorBatchRequest'
      responses:
        '202':
          description: Пачка обработана или поставлена в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BehaviorBatchResponse'
              example:
                accepted: 8
                duplicates: 1
                rejected: 0
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthenticated'
        '429':
          $ref: '#/components/responses/RateLimited'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session
      description: Случайный bearer token сессии MAX, действующий 24 часа; хранится на сервере только в виде hash.
  parameters:
...
    CityId:
      name: city_id
      in: query
      required: false
      schema:
        type: string
        format: uuid
      description: Город каталога. Если не передан, backend использует city_id из профиля пользователя; если его нет — VALIDATION_FAILED.
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
...
  responses:
    ValidationError:
      description: Ошибка валидации
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthenticated:
      description: Token отсутствует, истёк или невалиден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Превышен rate limit
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    CategorySlug:
      type: string
      enum:
      - concerts
      - cinema
      - theatre
      - standup
      - exhibitions
      - sports
      - food
      - parties
      - festivals
      - walks
      - other
...
    EventCard:
      type: object
      required:
      - id
      - title
      - category_slug
      - starts_at
      - timezone
      - date_label
      - venue_name
      - price_from_minor
      - currency
      - price_label
      - image_url
      - saved
      - reasons
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type:
          - string
          - 'null'
        category_slug:
          $ref: '#/components/schemas/CategorySlug'
        starts_at:
          type: string
          format: date-time
        other_occurrences_count:
          type: integer
          minimum: 0
          description: Количество других подходящих будущих сеансов этого же мероприятия.
          example: 3
        timezone:
          type: string
          example: Europe/Moscow
        date_label:
          type: string
          example: 19 сентября, 19:30
        venue_name:
          type: string
        latitude:
          type:
          - number
          - 'null'
          format: double
        longitude:
          type:
          - number
          - 'null'
          format: double
        distance_m:
          type:
          - integer
          - 'null'
        distance_label:
          type:
          - string
          - 'null'
        price_from_minor:
          type:
          - integer
          - 'null'
          minimum: 0
        currency:
          type: string
          enum:
          - RUB
        price_label:
          type: string
          example: от 1 800 ₽
        image_url:
          type:
          - string
          - 'null'
          format: uri
        saved:
          type: boolean
        reasons:
          type: array
          maxItems: 3
          items:
            $ref: '#/components/schemas/RecommendationReason'
    RecommendationReason:
      type: object
      required:
      - code
      - text
      properties:
        code:
          type: string
          enum:
          - time_fit
          - budget_fit
          - shared_category
          - nearby
          - profile_affinity
          - popular
        text:
          type: string
...
    HomeFeedResponse:
      type: object
      required:
      - feed_id
      - generated_at
      - sections
      properties:
        feed_id:
          type: string
          format: uuid
        generated_at:
          type: string
          format: date-time
        sections:
          type: array
          items:
            type: object
            required:
            - type
            - title
            - items
            properties:
              type:
                type: string
                enum:
                - hero
                - popular
                - for_you
                - nearby
              title:
                type: string
              items:
                type: array
                items:
                  $ref: '#/components/schemas/EventCard'
        active_room:
          anyOf:
          - $ref: '#/components/schemas/RoomSummary'
          - type: 'null'
...
    BehaviorBatchRequest:
      type: object
      additionalProperties: false
      required:
      - events
      properties:
        events:
          type: array
          minItems: 1
          maxItems: 100
          items:
            $ref: '#/components/schemas/BehaviorEvent'
    BehaviorEvent:
      type: object
      additionalProperties: false
      required:
      - client_event_id
      - type
      - event_version
      - occurred_at
      properties:
        client_event_id:
          type: string
          minLength: 1
          maxLength: 128
        type:
          type: string
          enum:
          - impression
          - open
          - share
          - app_opened
          - session_started
          - onboarding_started
          - onboarding_completed
          - feed_opened
          - event_impression
          - event_opened
          - search_performed
          - filters_opened
          - filters_applied
          - filters_reset
          - map_opened
          - map_marker_opened
          - room_creation_started
          - room_creation_failed
          - room_opened
          - invite_opened
          - invite_shared
          - invite_share_failed
          - invite_link_opened
          - room_join_started
          - room_join_failed
          - swipe_session_started
          - event_swipe_impression
          - match_shown
          - invite_shared
          - match_opened
          - swipe_pool_exhausted
          - client_error
          - client_performance
          - ticket_redirect_failed
        event_version:
          type: integer
          enum: [1]
        occurred_at:
          type: string
          format: date-time
        event_id:
          type:
          - string
          - 'null'
          format: uuid
        room_id:
          type:
          - string
          - 'null'
          format: uuid
        session_id:
          type:
          - string
          - 'null'
          format: uuid
        platform:
          type:
          - string
          - 'null'
          enum: [max_ios, max_android, max_web, max_desktop, browser, unknown, null]
        app_version:
          type:
          - string
          - 'null'
          maxLength: 64
        entry_point:
          type:
          - string
          - 'null'
          enum: [feed, home, search, map, saved, room_invite, room, deep_link, unknown, bot, direct, shared_event, recommendation, null]
        properties:
          type: object
          description: Whitelisted non-PII event properties. Raw search text, precise coordinates and secrets are prohibited.
          additionalProperties: true
        metadata:
          $ref: '#/components/schemas/BehaviorMetadata'
    BehaviorMetadata:
      type: object
      additionalProperties: false
      description: Только whitelisted metadata. PII, raw URLs, MAX init data, координаты и свободный текст запрещены.
      properties:
        surface:
          type: string
          maxLength: 32
        position:
          type: integer
          minimum: 0
        request_id:
          type: string
          maxLength: 128
    BehaviorBatchResponse:
      type: object
      required:
      - accepted
      - duplicates
      - rejected
      properties:
        accepted:
          type: integer
          minimum: 0
        duplicates:
          type: integer
          minimum: 0
        rejected:
          type: integer
          minimum: 0
...
    RoomState:
      type: string
      enum:
      - collecting_intents
      - ranking
      - voting
      - matched
      - exhausted
    RoomSummary:
      type: object
      required:
      - id
      - name
      - city_id
      - state
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        city_id:
          type: string
          format: uuid
        state:
          $ref: '#/components/schemas/RoomState'
    PublicParticipant:
      type: object
      required:
      - id
      - display_name
      - role
      - intent_ready
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        avatar_url:
          type:
          - string
          - 'null'
          format: uri
        role:
          type: string
          enum:
          - creator
          - participant
        intent_ready:
          type: boolean
          description: Подтверждение intent для текущего round; не раскрывает содержимое intent
...
    MatchSummary:
      type: object
      required:
      - id
      - room_id
      - event_id
      - matched_at
      - participants
      properties:
        id:
          type: string
          format: uuid
        room_id:
          type: string
          format: uuid
        event_id:
          type: string
          format: uuid
        matched_at:
          type: string
          format: date-time
        participants:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/PublicParticipant'
          description: Только безопасные display_name/avatar/role; MAX ID и intent не раскрываются
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          - request_id
          properties:
            code:
              type: string
              enum:
              - UNAUTHENTICATED
              - TOKEN_EXPIRED
              - FORBIDDEN
              - VALIDATION_FAILED
              - NOT_FOUND
              - CONFLICT
              - IDEMPOTENCY_CONFLICT
              - ACTIVE_ROOM_EXISTS
              - RATE_LIMITED
              - ROOM_FULL
              - INVITE_EXPIRED
              - ROOM_EXPIRED
              - ROOM_NOT_READY
              - POOL_NOT_READY
              - POOL_EXHAUSTED
              - STALE_POOL_VERSION
              - INTENT_LOCKED
              - VOTE_ALREADY_CAST
              - ALREADY_MATCHED
              - ROUND_LIMIT_REACHED
              - EVENT_UNAVAILABLE
              - TICKETS_UNAVAILABLE
              - INTERNAL
            message:
              type: string
            request_id:
              type: string
            details:
              type: object
              additionalProperties: true
              properties:
                current_pool_version:
                  type: integer
                  minimum: 1
                my_pool_finished:
                  type: boolean
                room_exhausted:
                  type: boolean
                field:
                  type: string
                fields:
                  type: array
                  items:
                    type: object
                    required:
                    - field
                    - reason
                    properties:
                      field:
                        type: string
                      reason:
                        type: string
                active_room:
                  anyOf:
                  - $ref: '#/components/schemas/RoomSummary'
                  - type: 'null'
                match:
                  anyOf:
                  - $ref: '#/components/schemas/MatchSummary'
                  - type: 'null'